Stylise the search and excludes ingredients fieldsets
