+++ src/App.css
@@ -1,7 +1,10 @@
+@import url("https://fonts.googleapis.com/css?family=Knewave&display=swap&subset=latin-ext");
+@import url("https://fonts.googleapis.com/css?family=Gloria+Hallelujah&display=swap");
 @import url('https://fonts.googleapis.com/css?family=Righteous&display=swap');
 @import url("https://fonts.googleapis.com/css?family=Gloria+Hallelujah&display=swap%22");
-@import url("https://fonts.googleapis.com/css?family=Sniglet:regular,800");
 @import url("https://fonts.googleapis.com/css?family=Gloria+Hallelujah|Indie+Flower&display=swap");
+@import url(https://fonts.googleapis.com/css?family=Sniglet:regular,800);
+@import url(https://fonts.googleapis.com/css?family=Gaegu:300,regular,700);
 /* @import url("https://fonts.google.com/specimen/Gloria+Hallelujah?selection.family=Gloria+Hallelujah&sidebar.open&preview.text=Salut+je+sais+pas+quoi+%C3%A9crire+howToBeFat.com+keep+calm+and+pet+your+cat&preview.text_type=custom"); */
 
 /* @import url("https://fonts.googleapis.com/css?family=Knewave&display=swap&subset=latin-ext%22"); */
