+++ src/components/RecipeSearch.css
@@ -1,4 +1,4 @@
-.ingredientSearch {
+/* .ingredientSearch {
     padding-left: 15px;
     padding-right: 15px;
     padding-top: 15px;
@@ -17,4 +17,4 @@ input {
     width: 100px;
     background-color: #00c2cb;
     margin-top: 15px;
-}
+} */
