+++ src/components/recipeSearch/Form/Form.js
@@ -12,20 +12,27 @@ import './Form.css'
 
 const Form = (props) => (
 
-  <div>
+  <form onSubmit={props.submitForm} className="FormContainer">
     <FormIngredients handleChange={props.handleChange} />
     <FormExcludes handleChange={props.handleChange} />
     <SpecialDiets handleChange={props.handleChange} />
     <Intolerables handleChange={props.handleChange} />
     <RangeCalories handleChange={props.handleChange} userCalories={props.userCalories} />
     <RangePrepTime handleChange={props.handleChange} userPrepTime={props.userPrepTime} />
-  </div>
+    {
+      props.errorRequest &&
+      <div style={{ padding: "1em", color: "red", "fontWeight": "bold" }}>
+        {props.errorRequest}
+      </div>
+    }
+    <div><input type="submit" value="Get recipe" className="button-recipe"></input></div>
+  </form>
 
 )
 
 Form.propTypes = {
   handleChange: PropTypes.func.isRequired,
-  onSubmit: PropTypes.func.isRequired,
+  submitForm: PropTypes.func.isRequired,
   errorRequest: PropTypes.bool.isRequired,
   userCalories: PropTypes.oneOfType([
     PropTypes.number.isRequired,
