+++ src/components/recipeSearch/Form/FormIngredients.css
@@ -1,0 +1,59 @@
+.ingredientFilter label {
+  text-align: center;
+  display: block;
+}
+
+.ingredientFilter input {
+  margin: auto;
+  position: relative;
+}
+
+.ingredientFilter .hideButton {
+  border-radius: 25%;
+  position: relative;
+  left: -5px;
+  font-weight: bold;
+  font-size: 1.2em;
+  border: 2px ridge #1c72a3;
+  padding: 0 .2em;
+  background-color: #1c72a3;
+  color: #ffcb05;
+}
+
+.ingredientFilter .hideButton:hover {
+  background-color: #ffcb05;
+  color: #1c72a3;
+}
+
+.ingredientFilter .hideButton:visited {
+  text-decoration: none;
+  color: #ffcb05;
+}
+
+.ingredientFilter .inputTexte {
+  text-align: center;
+  width: 60%;
+  border-radius: 25%;
+  padding: .2em;
+  border: 2px ridge #1c72a3;
+  font-size: 1.2em;
+  background-color: #ffcb05;
+  color: #1c72a3;
+  font-family: 'Sniglet', cursive;
+}
+
+.ingredientFilter .button-recipe {
+  background-color: #1c72a3;
+  color: #ffcb05;
+  margin-top: 1em;
+}
+
+.ingredientFilter .button-recipe:hover {
+  background-color: #ffcb05;
+  color: #1c72a3;
+}
+
+.ingredientFilter label {
+  font-family: 'Gaegu', cursive;
+  font-size: 1.2em;
+}
