+++ src/components/recipeSearch/Form/FormIngredients.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types'
 
 import './FormIngredients.css'
 
-const FormExcludes = (props) => {
+const FormIngredients = (props) => {
 
   const handleChange = props.handleChange
 
@@ -47,33 +47,33 @@ const FormExcludes = (props) => {
 
   return (
     <>
-      <fieldset className="ingredientSearch">
-        <legend> Exclude some ingredients </legend>
+      <fieldset className="ingredientFilter">
+        <legend>Search by ingredients</legend>
         {
           ingredientShow.ingredient1 &&
           <>
-            <label htmlFor='excludedIngredient1' id="labelIngredient1">Ingredient 1 <br />
+            <label htmlFor='ingredient1' id="labelIngredient1">Ingredient 1 <br />
               {!ingredientShow.ingredient2 &&
-                <input type="button" className="hideButton" onClick={handleHideInput} id="IngredientHideInput1" name="IngredientHideInput1" value="-" />
+                <input type="button" className="hideButton" onClick={handleHideInput} id="IngredientHideInput1" value="-" required />
               }
-              <input onChange={handleChange} id='excludedIngredient1' name='excludedIngredient1' type='text' pattern="[A-Za-z]+" />
+              <input onChange={handleChange} id='ingredient1' name='ingredient1' type='text' pattern="[A-Za-z]+" className="inputTexte" />
             </label>
           </>
         }
         {
           ingredientShow.ingredient2 &&
-          <label htmlFor='excludedIngredient2' id="labelIngredient2">Ingredient 2 <br />
+          <label htmlFor='ingredient2' id="labelIngredient2">Ingredient 2 <br />
             {!ingredientShow.ingredient3 &&
               <input type="button" className="hideButton" onClick={handleHideInput} id="IngredientHideInput2" value="-" />
             }
-            <input onChange={handleChange} id='excludedIngredient2' name='excludedIngredient2' type='text' pattern="[A-Za-z]+" />
+            <input onChange={handleChange} id='ingredient2' name="ingredient2" type='text' pattern="[A-Za-z]+" className="inputTexte" />
           </label>
         }
         {
           ingredientShow.ingredient3 &&
-          <label htmlFor='excludedIngredient3' id="labelIngredient3">Ingredient 3 <br />
+          <label htmlFor='ingredient3' id="labelIngredient3">Ingredient 3 <br />
             <input type="button" className="hideButton" onClick={handleHideInput} id="IngredientHideInput3" value="-" />
-            <input onChange={handleChange} id='excludedIngredient3' name="excludedIngredient3" type='text' pattern="[A-Za-z]+" />
+            <input onChange={handleChange} id='ingredient3' name="ingredient3" type='text' pattern="[A-Za-z]+" className="inputTexte" />
           </label>
         }
         {
@@ -85,8 +85,8 @@ const FormExcludes = (props) => {
   )
 }
 
-FormExcludes.propTypes = {
+FormIngredients.propTypes = {
   handleChange: PropTypes.func.isRequired
 }
 
-export default FormExcludes
+export default FormIngredients
