--- src/components/recipeSearch/FormIngredients.css
@@ -1,32 +1,0 @@
-.ingredientSearch label {
-  text-align: center;
-  display: block;
-  margin-bottom: .5em;
-}
-
-.ingredientSearch input {
-  margin: auto;
-  position: relative;
-}
-
-.ingredientSearch .hideButton {
-  border-radius: 25%;
-  position: relative;
-  left: -5px;
-  font-weight: bold;
-  font-size: 1.2em;
-  border: 2px ridge #1c72a3;
-  padding: 0 .2em;
-  background-color: #ffcb05;
-  color: #1c72a3;
-}
-
-.ingredientSearch .hideButton:hover {
-  background-color: #1c72a3;
-  color: #ffcb05;
-}
-
-.ingredientSearch .hideButton:visited {
-  text-decoration: none;
-  color: #1c72a3;
-}
